update mobile view
