+++ lib/contants.ts
@@ -4,12 +4,28 @@ import renew_logo from "@/public/images/renew_logo.jpg";
 import fiddle_logo from "@/public/images/fiddle_logo.jpg";
 import fitsnitch_logo from "@/public/images/fitsnitch_logo.png";
 import byu_logo from "@/public/images/byu_logo.png";
+import { SocialLink } from "../pages";
+import githublogo from "@/public/images/githublogo.png";
+import linkedInLogo from "@/public/images/linkedInLogo.png";
 
 export const NAME = "Renzo Caiña";
 export const JOB_TITLE = "Full Stack Software Developer";
 export const UNIVERSITY = "Brigham Young University";
 export const MAJOR = "Bachelor of Science in Computer Science";
 export const MINOR = "Minor in Business Management";
+
+export const socialLinks: SocialLink[] = [
+  {
+    name: "Github",
+    link: "https://github.com/rcaina",
+    img: githublogo,
+  },
+  {
+    name: "LinkedIn",
+    link: "https://www.linkedin.com/in/renzocaina/",
+    img: linkedInLogo,
+  },
+];
 
 export const EXPERIENCES = [
   {
